workbench_v2/requirements.txt
django>=5.0,<6.0
dj-database-url>=2.0
psycopg2-binary>=2.9
google-generativeai>=0.8
gunicorn>=22.0
whitenoise>=6.0
Pillow>=10.0
